Refactored partial class usage as it was an ugly and bad quickfix - Added Services that contain the methods commonly used by the modules - moved pickup modules to a separate subfolder in modules - moved the service locator to the commands library since it will never be used in the main console app
